+++ Constraint/ConstraintParser.h
@@ -25,6 +25,7 @@ int create_new_constraint(const shared_ptr<DomainBase>&, istringstream&);
 void new_embed(unique_ptr<Constraint>&, istringstream&, unsigned);
 void new_fixedlength(unique_ptr<Constraint>&, istringstream&, unsigned);
 void new_minimumgap(unique_ptr<Constraint>&, istringstream&, unsigned);
+void new_maximumgap(unique_ptr<Constraint>&, istringstream&, unsigned);
 void new_mpc(unique_ptr<Constraint>&, istringstream&);
 void new_particlecollision2d(unique_ptr<Constraint>&, istringstream&);
 void new_particlecollision3d(unique_ptr<Constraint>&, istringstream&);