+++ Constraint/FixedLength.h
@@ -35,13 +35,12 @@
 #include <Constraint/Constraint.h>
 
 class FixedLength : public Constraint {
-	const bool inequal;
-	const double min_gap;
-
 	vec coor;
+protected:
+	const bool min_bound = false, max_bound = false;
+	const double min_gap = 0., max_gap = 0.;
 public:
 	FixedLength(unsigned, unsigned, unsigned, unsigned, uvec&&);
-	FixedLength(unsigned, unsigned, unsigned, unsigned, double, uvec&&);
 
 	int initialize(const shared_ptr<DomainBase>&) override;
 
@@ -56,6 +55,11 @@ public:
 class MinimumGap final : public FixedLength {
 public:
 	MinimumGap(unsigned, unsigned, unsigned, unsigned, double, uvec&&);
+};
+
+class MaximumGap final : public FixedLength {
+public:
+	MaximumGap(unsigned, unsigned, unsigned, unsigned, double, uvec&&);
 };
 
 #endif